+++ pkg/frontend/frontend_test.go
@@ -35,6 +35,7 @@ import (
 	"github.com/grafana/mimir/pkg/frontend/transport"
 	"github.com/grafana/mimir/pkg/frontend/v1/frontendv1pb"
 	querier_worker "github.com/grafana/mimir/pkg/querier/worker"
+	"github.com/grafana/mimir/pkg/scheduler/schedulerdiscovery"
 )
 
 const (
@@ -279,6 +280,11 @@ func defaultFrontendConfig() CombinedFrontendConfig {
 	flagext.DefaultValues(&config.Handler)
 	flagext.DefaultValues(&config.FrontendV1)
 	flagext.DefaultValues(&config.FrontendV2)
+
+	querySchedulerDiscoveryConfig := schedulerdiscovery.Config{}
+	flagext.DefaultValues(&querySchedulerDiscoveryConfig)
+	config.FrontendV2.QuerySchedulerDiscovery = querySchedulerDiscoveryConfig
+
 	return config
 }